requirements.txt
streamlit>=1.28.0
requests>=2.31.0
typing-extensions>=4.5.0
PyMuPDF
sentence-transformers
faiss-cpu